backend/requirements.txt
# VectorSurfer 2.0 Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0

# Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Weaviate
weaviate-client>=4.0.0
python-dotenv>=1.0.0

# OpenAI (for embeddings + RAG features)
openai>=1.0.0


# [추가] For async batch processing in sync context
nest-asyncio>=1.5.0

# [BYOD] Database & Auth
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
cachetools>=5.3.0
cryptography>=42.0.0

# [GitHub Integration] Async HTTP client
httpx>=0.27.0
